+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This is a list of manually mantained changes and updates for each version.
 
 ## Changed
 
+### Router response bug fix for non-gzipped encoding
+
+### dependency gracelog version updated to 0.6.0
+
 ### Render parser for if has been improved
 
 ## Deprecated